Plex-Meta-Manager/requirements.txt
PlexAPI
tmdbapis
tmdbv3api
python-dotenv
tvdb_v4_official
ruamel.yaml
requests
pathvalidate
